ALEPH-66 #comment moved logging schema into management schema
